fix: 당첨금 자료형 수정 및 오타 수정

+++ src/main/java/lotto/view/OutputView.java
@@ -58,7 +58,7 @@ public class OutputView {
         System.out.println(message);
     }
 
-    private static String formatWinnings(int winnings) {
+    private static String formatWinnings(long winnings) {
         return String.format(WINNINGS_FORMAT, winnings);
     }
